recovery: Make an sdcard image that just boots the recovery.fit
It is sometimes nice to have a simple test payload.
So use the recovery assets we build for this.
A separate sdcard image is built for this.
It should/can be used with the recovery firmware as ImageA.bin

Since the recovery firmware knows if it is in recovery mode or not
it can be used that way.  The recovery.fit is also pretty generic
as it just boots and waits to be told what to do.
So all we need is a new script that loads and boots the fit image.

* build the recovery-boot-script
* rename do_sdcard to do_one_sdcard and parameterize it.
* parameterize do_sudo_sdcard
* in main section, pass any extra command line arguments to subscripts
* rename recovery-sdcard.img to recovery-update-sdcard.img
* In new do_sdcard build both sdcard versions

Signed-off-by: Bill Mills <bill.mills@linaro.org>