kira-deploy: delete kira-update-recovery mode
This mode could not update the recovery-boot.bin mtd partitions from the
kernel because there are marked read-only.  We could make a custom DTB for
the fit that made them R/W but making an sdcard was needed any way.
So delete this mode.  It can always be brought back if needed.

We now have the recovery-update-sdcard.img to update the on board recovery
assets.  This sdcard can be used manually or be deployed as a lava job.
This will enforce that we have a know good normal mode before it updates
the recovery mode.

The sdcard image should be paired with know good firmware image.
recovery-boot.bin can be used for this if called ImageA.bin.
It knows when it is running from the recovery slot or not.
When not, it will boot the sdcard.

Signed-off-by: Bill Mills <bill.mills@linaro.org>