recovery-zcu102: Make SD images from PetaLinux 2022.2
recovery-zcu102:
I have re-built the Xilinx 2022.2 petalinux release using
the for-ref/xilinx-vendor.sh procedure of openamp-ci-builds.
https://github.com/wmamills/openamp-ci-builds/tree/main/for-ref

The boot firmware, kernel and user-space image are checked in here for
conveyance.

The "./mk-sd-images.sh all"  command will create 3 sd card images.
A tftp only SD card of 64MB
A vfat only boot with initramfs of 128MB.
And a vfat + ext4 rootfs of 1GB.

sudo is required to build the sd images.

Signed-off-by: Bill Mills <bill.mills@linaro.org>