zcu102 Lava: check in working job and log
This version of the job use my customized SD card image.
The image has a blank root and petalinux password.
I use the vfat+ext4 version here so it could be used with lava overlays etc
if someone wished to.  The SD card is 1GB with 650MB free.

The vfat-initrd version works just the same and deploys faster but may be
harder to expand or customize.

Signed-off-by: Bill Mills <bill.mills@linaro.org>