kria-deploy.sh: support newer scp in openssh-client 9
Newer versions of openssh-client use sftp-server remote command for scp.
Dropbear used in the Kria recovery image does not support this.
So we need to use the -O option to scp when available.

This support is required in debian 12 bookworm and Ubuntu 24.04.

Test with docker images for
debian 10 11 and 12
ubuntu 20.04 22.04 and 24.04 (24.04 is beta at this time)

This patch also completely disables know_host file usage.
Before the know_host file was updated each run with a new host key.
Now know_host processing and checking is disabled for each ssh
& scp command.  This allows the user to not have a ~/.ssh dir at all.

Signed-off-by: Bill Mills <bill.mills@linaro.org>